requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==3.2.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
easyocr==1.7.1
execnet==2.0.2
filelock==3.13.1
fsspec==2024.2.0
func_timeout==4.3.5
greenlet==3.0.3
h11==0.14.0
idna==3.6
imageio==2.34.0
iniconfig==2.0.0
Jinja2==3.1.3
lazy_loader==0.3
loguru==0.7.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
playwright==1.41.2
pluggy==1.4.0
pyclipper==1.3.0.post5
pycparser==2.21
pyee==11.0.1
pyOpenSSL==24.0.0
PySocks==1.7.1
pytest==8.0.2
pytest-assume==2.4.3
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-bidi==0.4.2
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
selenium==4.18.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.12
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
